Prepare list games and libraries

+++ lib/sdl/include/sdl/IButton.hpp
@@ -17,7 +17,7 @@ namespace arc
             virtual ~IButton() = default;
 
             virtual void setPosition(size_t x, size_t y) = 0;
-            virtual bool isMouseHover(size_t x, size_t y) const = 0;
+            virtual bool isMouseHover(int x, int y) const = 0;
             virtual void display(SDL_Renderer *window) = 0;
     };
 } // namespace arc
